テストのやつをそれっぽく実装
