+++ package.json
@@ -6,6 +6,7 @@
     "discord.js": "^14.3.0",
     "dotenv": "^16.0.2",
     "enquirer": "^2.3.6",
+    "moment": "^2.29.4",
     "node-cron": "^3.0.2"
   },
   "name": "nitkc-main",
